+++ contracts/libraries/TickBitMap.sol
@@ -14,7 +14,7 @@ import '../libraries/TickMath.sol';
 library TickBitMap {
     // computes the position in the uint256 array where the initialized state for a tick lives
     // bitPos is the 0 indexed position in the word from most to least significant where the flag is set
-    function position(int16 tick) private pure returns (uint256 wordPos, uint256 bitPos) {
+    function position(int32 tick) private pure returns (uint256 wordPos, uint256 bitPos) {
         require(tick >= TickMath.MIN_TICK, 'TickBitMap::position: tick must be greater than or equal to MIN_TICK');
         require(tick <= TickMath.MAX_TICK, 'TickBitMap::position: tick must be less than or equal to MAX_TICK');
         // this subtraction is safe because tick - TickMath.MIN_TICK is at most 7351 * 2 which fits within int16
@@ -24,14 +24,14 @@ library TickBitMap {
     }
 
     // returns whether the given tick is initialized
-    function isInitialized(uint256[58] storage self, int16 tick) internal view returns (bool) {
+    function isInitialized(uint256[58] storage self, int32 tick) internal view returns (bool) {
         (uint256 wordPos, uint256 bitPos) = position(tick);
         uint256 mask = uint256(1) << bitPos;
         return self[wordPos] & mask != 0;
     }
 
     // flips the tick from uninitialized to initialized, or vice versa
-    function flipTick(uint256[58] storage self, int16 tick) internal {
+    function flipTick(uint256[58] storage self, int32 tick) internal {
         (uint256 wordPos, uint256 bitPos) = position(tick);
         uint256 mask = uint256(1) << bitPos;
         self[wordPos] ^= mask;
@@ -41,9 +41,9 @@ library TickBitMap {
     // or greater than this tick
     function nextInitializedTickWithinOneWord(
         uint256[58] storage self,
-        int16 tick,
+        int32 tick,
         bool lte
-    ) internal view returns (int16 next, bool initialized) {
+    ) internal view returns (int32 next, bool initialized) {
         if (lte) {
             (uint256 wordPos, uint256 bitPos) = position(tick);
             uint256 word = self[wordPos];
@@ -52,9 +52,9 @@ library TickBitMap {
             uint256 masked = word & mask;
 
             // there are no initialized ticks to the left or at of the current tick, return the leftmost in the word
-            if (masked == 0) return (tick - int16(255 - bitPos), false);
+            if (masked == 0) return (tick - int32(255 - bitPos), false);
 
-            return (tick + (int16(bitPos) - int16(BitMath.leastSignificantBit(masked))), true);
+            return (tick + (int32(bitPos) - int32(BitMath.leastSignificantBit(masked))), true);
         } else {
             // start from the word of the next tick, since the current tick state doesn't matter
             (uint256 wordPos, uint256 bitPos) = position(tick + 1);
@@ -64,18 +64,18 @@ library TickBitMap {
             uint256 masked = word & mask;
 
             // there are no initialized ticks to the right of the current tick, just return the rightmost in the word
-            if (masked == 0) return (tick + 1 + int16(bitPos), false);
+            if (masked == 0) return (tick + 1 + int32(bitPos), false);
 
-            return ((tick + 1) + (int16(bitPos) - int16(BitMath.mostSignificantBit(masked))), true);
+            return ((tick + 1) + (int32(bitPos) - int32(BitMath.mostSignificantBit(masked))), true);
         }
     }
 
     // same as above, but iterates until it finds the next initialized tick
     function nextInitializedTick(
         uint256[58] storage self,
-        int16 tick,
+        int32 tick,
         bool lte
-    ) internal view returns (int16 next) {
+    ) internal view returns (int32 next) {
         bool initialized;
         next = tick;
         if (lte) {
