+++ contracts/test/TickBitMapEchidnaTest.sol
@@ -8,17 +8,17 @@ contract TickBitMapEchidnaTest {
 
     uint256[58] public bitmap;
 
-    function flipTick(int16 tick) public {
+    function flipTick(int32 tick) public {
         bitmap.flipTick(tick);
     }
 
-    function checkNextInitializedTickWithinOneWordInvariants(int16 tick, bool lte) public view {
-        (int16 next, bool initialized) = bitmap.nextInitializedTickWithinOneWord(tick, lte);
+    function checkNextInitializedTickWithinOneWordInvariants(int32 tick, bool lte) public view {
+        (int32 next, bool initialized) = bitmap.nextInitializedTickWithinOneWord(tick, lte);
         if (lte) {
             assert(next <= tick);
             assert(tick - next < 256);
             // all the ticks between the input tick and the next tick should be uninitialized
-            for (int16 i = tick - 1; i > next; i--) {
+            for (int32 i = tick - 1; i > next; i--) {
                 assert(!bitmap.isInitialized(i));
             }
             assert(bitmap.isInitialized(next) == initialized);
@@ -26,26 +26,26 @@ contract TickBitMapEchidnaTest {
             assert(next > tick);
             assert(next - tick <= 256);
             // all the ticks between the input tick and the next tick should be uninitialized
-            for (int16 i = tick + 1; i < next; i++) {
+            for (int32 i = tick + 1; i < next; i++) {
                 assert(!bitmap.isInitialized(i));
             }
             assert(bitmap.isInitialized(next) == initialized);
         }
     }
 
-    function checkNextInitializedTickInvariants(int16 tick, bool lte) public view {
-        int16 next = bitmap.nextInitializedTick(tick, lte);
+    function checkNextInitializedTickInvariants(int32 tick, bool lte) public view {
+        int32 next = bitmap.nextInitializedTick(tick, lte);
         if (lte) {
             assert(next <= tick);
             // all the ticks between the input tick and the next tick should be uninitialized
-            for (int16 i = tick - 1; i > next; i--) {
+            for (int32 i = tick - 1; i > next; i--) {
                 assert(!bitmap.isInitialized(i));
             }
             assert(bitmap.isInitialized(next));
         } else {
             assert(next > tick);
             // all the ticks between the input tick and the next tick should be uninitialized
-            for (int16 i = tick + 1; i < next; i++) {
+            for (int32 i = tick + 1; i < next; i++) {
                 assert(!bitmap.isInitialized(i));
             }
             assert(bitmap.isInitialized(next));
